Moved shifting code from Scan to Shift
